lint: fix file_length and function_body_length in SettingsView — split RemovalAlertModifier to own file, extract sheet modifiers
